backend_python/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
apscheduler
sqlalchemy
pymysql
passlib[bcrypt]
python-jose
PyMuPDF
farm-haystack==1.26.4
pydantic<2.0,>=1.10.0
typing-inspection==0.4.0
transformers>=4.30.0
torch>=1.13.0
selenium
undetected_chromedriver
transformers 